avoid Long in test

+++ test/shared/src/main/scala/scalaprops/TestsDefinitions.scala
@@ -18,7 +18,7 @@ object TestsDefinitions {
 
   sealed trait Base
   case class BaseIS(i: Int, s: String) extends Base
-  case class BaseDB(d: Long, b: Boolean) extends Base
+  case class BaseDB(d: Byte, b: Boolean) extends Base
   case class BaseLast(c: Simple) extends Base
 
   case class CCWithSingleton(i: Int, s: Witness.`"aa"`.T)
@@ -58,7 +58,7 @@ object TestsDefinitions {
   sealed case class B(i: Int, s: String) extends A
   case object C extends A
   sealed trait D extends A
-  final case class E(a: Long, b: Option[Int]) extends D
+  final case class E(a: Byte, b: Option[Int]) extends D
   case object F extends D
   sealed abstract class Foo extends D
   case object Baz extends Foo
@@ -79,7 +79,7 @@ object TestsDefinitions {
 
     sealed trait BaseNoGen
     case class BaseNoGenIS(i: Int, s: String) extends BaseNoGen
-    case class BaseNoGenDB(d: Long, b: Boolean) extends BaseNoGen
+    case class BaseNoGenDB(d: Byte, b: Boolean) extends BaseNoGen
     case class BaseNoGenN(n: NoGenitraryType) extends BaseNoGen
   }
